# server/docs/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "tubu.io API Documentation - ",
    "version": "0.2.0",
    "title": "TUBU.IO API",
    "contact": {
      "email": "dev@tubu.io"
    }
  },
...
  "paths": {
...
    "/api/network/{networkID}/contract": {
      "post": {
        "tags": ["Contract"],
        "summary": "Deploy new contract",
        "description": "",
        "operationId": "deployContract",
        "parameters": [
          {
            "$ref": "#/components/parameters/networkID"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/BasicContract"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully deployed contract",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "contract_address": {
                      "type": "string"
                    },
                    "short_id": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "abi": {
                      "type": "string"
                    },
                    "owner_address": {
                      "type": "string"
                    },
                    "user_id": {
                      "type": "integer"
                    },
                    "tag": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request with error message",
            "content": {
              "application/json": {
                "schema": {
                  "error": "object",
                  "message": "string"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      },
      "get": {
        "tags": ["Contract"],
        "summary": "Get current user's all contracts",
        "description": "",
        "operationId": "getContracts",
        "parameters": [
          {
            "$ref": "#/components/parameters/networkID"
          },
          {
            "$ref": "#/components/parameters/pageFilter"
          },
          {
            "$ref": "#/components/parameters/pageSizeFilter"
          },
          {
            "in": "query",
            "name": "appID",
            "description": "Application ID",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "contract_address": {
                          "type": "string"
                        },
                        "short_id": {
                          "type": "string"
                        },
                        "updated_at": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "abi": {
                          "type": "string"
                        },
                        "owner_address": {
                          "type": "string"
                        },
                        "user_id": {
                          "type": "integer"
                        },
                        "tag": {
                          "type": "string"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "totalCount": {
                          "type": "integer"
                        },
                        "totalPage": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "error": "object",
            "message": "string"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      }
    },
    "/api/network/{networkID}/contract/{shortID}": {
      "get": {
        "tags": ["Contract"],
        "summary": "Get current user's contract's latest version",
        "description": "",
        "operationId": "getContractLatestVersion",
        "parameters": [
          {
            "$ref": "#/components/parameters/shortID"
          },
          {
            "$ref": "#/components/parameters/networkID"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "description": "Get Tagged Contract Version (Optional) (Default Latest Version)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "version": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string"
                            },
                            "contract_address": {
                              "type": "string"
                            },
                            "tag": {
                              "type": "string"
                            },
                            "abi": {
                              "type": "object"
                            }
                          }
                        },
                        "versionList": {
                          "id": { "type": "integer" },
                          "description": { "type": "string" },
                          "tag": { "type": "string" }
                        },
                        "contract": {
                          "application_id": { "type": "integer" },
                          "name": { "type": "string" },
                          "description": { "type": "string" },
                          "short_id": { "type": "string" }
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "error": "object",
            "message": "string"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      },
      "delete": {
        "tags": ["Contract"],
        "summary": "Delete the contract with given shortID",
        "description": "",
        "operationId": "deleteContract",
        "parameters": [
          {
            "$ref": "#/components/parameters/shortID"
          },
          {
            "$ref": "#/components/parameters/networkID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "short_id": {
                          "type": "string"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "error": "object",
            "message": "string"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      },
      "put": {
        "tags": ["Contract"],
        "summary": "Update the name and description of the contract's latest version",
        "description": "",
        "operationId": "updateContractLatestVersionInfo",
        "parameters": [
          {
            "$ref": "#/components/parameters/shortID"
          },
          {
            "$ref": "#/components/parameters/networkID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "New Best Contract"
                  },
                  "description": {
                    "type": "string",
                    "example": "This is the most important updated contract description"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "400": {
            "error": "object",
            "message": "string"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      }
    },
...
  },
  "security": [
    {
      "UserToken": []
    }
  ],
  "components": {
    "securitySchemes": {
      "UserToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
...
      "BasicContract": {
        "type": "object",
        "required": ["name", "application_id"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Contract Name",
            "example": "sample contract"
          },
          "application_id": {
            "type": "integer",
            "description": "Contract Application",
            "example": 1
          },
          "description": {
            "type": "string",
            "example": "The contract is to avoid bad things"
          },
          "tag": {
            "type": "string",
            "description": "The tag for the first deployment of the contract",
            "default": "v1.0"
          },
          "short_id": {
            "type": "string",
            "description": "The shortID of the contract (to deploy a new version)(optional)"
          },
          "files": {
            "type": "array",
            "items": {
              "type": "file"
            }
          },
          "args": {
            "type": "string"
          }
        },
        "xml": {
          "name": "BasicContract"
        }
      },
...
    },
...
    "parameters": {
      "networkID": {
        "in": "path",
        "name": "networkID",
        "description": "Network ID",
        "schema": {
          "minimum": 1,
          "default": 1,
          "type": "integer"
        }
      },
...
      "shortID": {
        "in": "path",
        "name": "shortID",
        "description": "Short ID for contract",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "pageFilter": {
        "in": "query",
        "name": "page",
        "description": "Active Page Number",
        "schema": {
          "minimum": 1,
          "default": 1,
          "type": "integer"
        }
      },
      "pageSizeFilter": {
        "in": "query",
        "name": "pageSize",
        "description": "Item count on a page",
        "schema": {
          "minimum": 1,
          "default": 20,
          "type": "integer"
        }
      },
...
    }
  }
}